处理macos下编译兼容性

+++ crates/media-agent/src/heartbeat.rs
@@ -120,8 +120,8 @@ fn sample_disk_percent(path: &str) -> Option<f64> {
     }
 
     let stat = unsafe { stat.assume_init() };
-    let total = stat.f_blocks.saturating_mul(stat.f_frsize);
-    let free = stat.f_bavail.saturating_mul(stat.f_frsize);
+    let total = (stat.f_blocks as u64).saturating_mul(stat.f_frsize as u64);
+    let free = (stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64);
     if total == 0 {
         return Some(0.0);
     }
